rename head_object ==> get_object_metadata
